feature: renaming files based on the original file name; fixes #25700
This commit adds a new `namepart` called `originalfilename`, that can be used to create a naming pattern that is based on the original file name. After every renaming, the plugin saves the new filename with a reference to the original filename in the process properties as json. This ensures, that multiple invocations result in the same output and that the original filename can always be references, even if overwritten meanwhile.